requirements.txt
setuptools
django
djangorestframework
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
drf-yasg